+++ src/state.py
@@ -37,25 +37,111 @@ class BackendState:
 
 
 @dataclass(frozen=True)
+class HealthCheck:
+    """Charm state that contains the health check configuration.
+
+    Attributes:
+        path: The path to use for server health checks.
+        port: The port to use for http-check.
+        interval: Interval between health checks in seconds.
+        rise: Number of successful health checks before server is considered up.
+        fall: Number of failed health checks before server is considered down.
+    """
+
+    path: str | None
+    port: int | None
+    interval: int | None = Field(gt=1)
+    rise: int | None = Field(gt=1)
+    fall: int | None = Field(gt=1)
+
+    @classmethod
+    def from_charm(cls, charm: ops.CharmBase) -> "HealthCheck":
+        """Create an HealthCheck class from a charm instance.
+
+        Args:
+            charm: the ingress-configurator charm.
+
+        Returns:
+            HealthCheck: instance of the health check component.
+        """
+        interval = (
+            cast(int, charm.config.get("check-interval"))
+            if charm.config.get("check-interval")
+            else None
+        )
+        rise = (
+            cast(int, charm.config.get("check-rise")) if charm.config.get("check-rise") else None
+        )
+        fall = (
+            cast(int, charm.config.get("check-fall")) if charm.config.get("check-fall") else None
+        )
+        path = (
+            cast(str, charm.config.get("check-path")) if charm.config.get("check-path") else None
+        )
+        port = (
+            cast(int, charm.config.get("check-port")) if charm.config.get("check-port") else None
+        )
+        return cls(interval=interval, rise=rise, fall=fall, path=path, port=port)
+
+
+@dataclass(frozen=True)
+class Retry:
+    """Charm state that contains the retry configuration.
+
+    Attributes:
+        count: Number of times to retry failed requests.
+        interval: Interval between retries in seconds.
+        redispatch: Whether to redispatch failed requests to another server.
+    """
+
+    count: int | None = Field(gt=0)
+    interval: int | None = Field(gt=0)
+    redispatch: bool | None = False
+
+    @classmethod
+    def from_charm(cls, charm: ops.CharmBase) -> "Retry":
+        """Create an Retry class from a charm instance.
+
+        Args:
+            charm: the ingress-configurator charm.
+
+        Returns:
+            Retry: instance of the retry component.
+        """
+        count = (
+            cast(int, charm.config.get("retry-count")) if charm.config.get("retry-count") else None
+        )
+        interval = (
+            cast(int, charm.config.get("retry-interval"))
+            if charm.config.get("retry-interval")
+            else None
+        )
+        redispatch = (
+            cast(bool, charm.config.get("retry-redispatch"))
+            if charm.config.get("retry-redispatch")
+            else None
+        )
+        return cls(count=count, interval=interval, redispatch=redispatch)
+
+
+@dataclass(frozen=True)
 class State:
     """Charm state that contains the configuration.
 
     Attributes:
         backend_addresses: Configured list of backend ip addresses.
         backend_ports: Configured list of backend ports.
+        check: Health check configuration.
+        retry: Retry configuration.
         service: The service name.
-        retry_count: Number of times to retry failed requests.
-        retry_interval: Interval between retries in seconds.
-        retry_redispatch: Whether to redispatch failed requests to another server.
         paths: List of URL paths to route to the service.
         subdomains: List of subdomains to route to the service.
     """
 
     _backend_state: BackendState
+    check: HealthCheck
+    retry: Retry
     service: str = Field(..., min_length=1)
-    retry_count: int | None = Field(gt=0)
-    retry_interval: int | None = Field(gt=0)
-    retry_redispatch: bool | None = False
     paths: list[Annotated[str, BeforeValidator(value_has_valid_characters)]] = Field(default=[])
     subdomains: list[Annotated[str, BeforeValidator(value_has_valid_characters)]] = Field(
         default=[]
@@ -72,7 +158,6 @@ class State:
         return self._backend_state.backend_ports
 
     @classmethod
-    # pylint: disable=too-many-locals
     def from_charm(cls, charm: ops.CharmBase, ingress_data: IngressRequirerData | None) -> "State":
         """Create an State class from a charm instance.
 
@@ -108,19 +193,6 @@ class State:
             if charm.config.get("paths")
             else []
         )
-        retry_count = (
-            cast(int, charm.config.get("retry-count")) if charm.config.get("retry-count") else None
-        )
-        retry_interval = (
-            cast(int, charm.config.get("retry-interval"))
-            if charm.config.get("retry-interval")
-            else None
-        )
-        retry_redispatch = (
-            cast(bool, charm.config.get("retry-redispatch"))
-            if charm.config.get("retry-redispatch")
-            else None
-        )
         subdomains = (
             cast(str, charm.config.get("subdomains")).split(CHARM_CONFIG_DELIMITER)
             if charm.config.get("subdomains")
@@ -137,9 +209,8 @@ class State:
             return cls(
                 _backend_state=BackendState(backend_addresses, backend_ports),
                 paths=paths,
-                retry_count=retry_count,
-                retry_interval=retry_interval,
-                retry_redispatch=retry_redispatch,
+                check=HealthCheck.from_charm(charm),
+                retry=Retry.from_charm(charm),
                 service=f"{charm.model.name}-{charm.app.name}",
                 subdomains=subdomains,
             )