+++ tests/unit/test_state.py
@@ -24,6 +24,11 @@ def test_adapter_state_from_charm():
     """
     charm = Mock(CharmBase)
     charm.config = {
+        "check-interval": 20,
+        "check-rise": 3,
+        "check-fall": 4,
+        "check-path": "/health",
+        "check-port": 8080,
         "retry-count": 1,
         "retry-interval": 10,
         "retry-redispatch": True,
@@ -35,13 +40,19 @@ def test_adapter_state_from_charm():
         units=[IngressRequirerUnitData(host="sample.host", ip="127.0.0.1")],
     )
     charm_state = state.State.from_charm(charm, ingress_relation_data)
+
     assert [str(address) for address in charm_state.backend_addresses] == [
         ingress_relation_data.units[0].ip
     ]
     assert charm_state.backend_ports == [ingress_relation_data.app.port]
-    assert charm_state.retry_count == charm.config.get("retry-count")
-    assert charm_state.retry_interval == charm.config.get("retry-interval")
-    assert charm_state.retry_redispatch == charm.config.get("retry-redispatch")
+    assert charm_state.check.interval == charm.config.get("check-interval")
+    assert charm_state.check.rise == charm.config.get("check-rise")
+    assert charm_state.check.fall == charm.config.get("check-fall")
+    assert charm_state.check.path == charm.config.get("check-path")
+    assert charm_state.check.port == charm.config.get("check-port")
+    assert charm_state.retry.count == charm.config.get("retry-count")
+    assert charm_state.retry.interval == charm.config.get("retry-interval")
+    assert charm_state.retry.redispatch == charm.config.get("retry-redispatch")
     assert charm_state.paths == charm.config.get("paths").split(",")
     assert charm_state.subdomains == charm.config.get("subdomains").split(",")
 
@@ -67,9 +78,9 @@ def test_integrator_state_from_charm():
     assert [str(port) for port in charm_state.backend_ports] == charm.config.get(
         "backend-ports"
     ).split(",")
-    assert charm_state.retry_count == charm.config.get("retry-count")
-    assert charm_state.retry_interval == charm.config.get("retry-interval")
-    assert charm_state.retry_redispatch == charm.config.get("retry-redispatch")
+    assert charm_state.retry.count == charm.config.get("retry-count")
+    assert charm_state.retry.interval == charm.config.get("retry-interval")
+    assert charm_state.retry.redispatch == charm.config.get("retry-redispatch")
 
 
 def test_state_from_charm_no_backend():
@@ -125,6 +136,48 @@ def test_state_from_charm_invalid_port():
     charm.config = {
         "backend-addresses": "127.0.0.1,127.0.0.2",
         "backend-ports": "99999",
+    }
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
+def test_state_from_charm_invalid_check_interval():
+    """
+    arrange: mock a charm with backend address and without check-interval configuration
+    act: instantiate a State
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {
+        "check-interval": -1,
+    }
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
+def test_state_from_charm_invalid_check_rise():
+    """
+    arrange: mock a charm with backend address and without check-rise configuration
+    act: instantiate a State
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {
+        "check-rise": 0,
+    }
+    with pytest.raises(state.InvalidStateError):
+        state.State.from_charm(charm, None)
+
+
+def test_state_from_charm_invalid_check_fall():
+    """
+    arrange: mock a charm with backend address and without check-fall configuration
+    act: instantiate a State
+    assert: a InvalidStateError is raised
+    """
+    charm = Mock(CharmBase)
+    charm.config = {
+        "check-fall": 0,
     }
     with pytest.raises(state.InvalidStateError):
         state.State.from_charm(charm, None)